feat: prepare repository for public release v1.0.0
- Add community health files: SECURITY.md, CHANGELOG.md, CODE_OF_CONDUCT.md

- Add GitHub issue/PR templates and CODEOWNERS

- Add dependabot.yml for automated cargo + Actions updates

- Fix all URLs to N1KH1LT0X1N/WinSweep

- Fix .gitignore: track Cargo.lock, remove duplicate dist/

- Fix CI: install MinGW, add gnu target, fix artifact paths

- Update release.yml: action-gh-release v2

- Move plan.md to docs/archive/

- Commit 33 new package managers, browser cache, scheduler, notifications, locales, installer, scripts, and all updated source files

+++ crates/winsweep-core/src/audit_logger.rs
@@ -9,7 +9,7 @@ use serde::{Deserialize, Serialize};
 use std::fs::OpenOptions;
 use std::io::Write;
 use std::path::PathBuf;
-use tracing::{debug, error, info, warn};
+use tracing::{debug, info};
 use uuid::Uuid;
 
 /// Audit log entry
@@ -361,7 +361,7 @@ impl AuditLogger {
         let metadata = std::fs::metadata(&self.log_file)?;
         let size_mb = metadata.len() / (1024 * 1024);
 
-        if size_mb > max_size_mb {
+        if size_mb > max_size_mb as u64 {
             let timestamp = Utc::now().format("%Y%m%d_%H%M%S");
             let rotated_path = self.log_file.with_extension(format!("log.{}", timestamp));
